Simplify kmap saver to use library interfaces
We want RME code to be de-atached from flatbuffers, kmap and the kmap library. For that we've introduced a interface based library and we've implemented the interfaces in RME side (some of them were just incorporated to RME basic entities, like position).

That way we could remove a lot of boiler plate check, simplify the logics in the save and delegate some validations to the new domain objects.