improve: remove boosts library and add json manual library
Removed boosts (replaced with equivalent on std 14/17/20)
Removed manual json library implementation using boost and added nlohmann-json

So two additional libraries will be needed to compile:
asio,
nlohmann-json

I also removed OTGZ support by default and disabled it, making it optional (needs to be enabled). Anyone who wants to use it will need to activate this, I think it's ideal since few should use it, so a majority don't use it, which is fairer that those who need to activate it to use it than those who don't need to have this in their program.